added content provider to lesson 10

+++ lesson10/src/main/java/com/example/lesson10/entities/MyNote.java
@@ -10,7 +10,7 @@ import android.support.annotation.NonNull;
 public class MyNote implements Parcelable {
 
     @PrimaryKey(autoGenerate = true)
-    private int id;
+    private long id;
     private String name;
     private String date;
     private String content;
@@ -43,11 +43,11 @@ public class MyNote implements Parcelable {
         }
     };
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -82,7 +82,7 @@ public class MyNote implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeString(name);
         dest.writeString(date);
         dest.writeString(content);
